PRO-3070: Reinstate old Domicile page and create new stand-alone Domicile page
