+++ app/resources/en/translation/deceased/domicile.json
@@ -1,16 +1,15 @@
 {
   "title": "Deceased domicile",
-  "question": "Did the person who died live permanently in England or Wales, or intend to return here?",
-  "hintText1": "You can only apply for probate online if the person who died was domiciled in England or Wales at the time of their death.",
-  "hintText2": "You can read more about <a href=\"https://www.gov.uk/inheritance-tax/when-someone-living-outside-the-uk-dies\">domicile</a>.",
-  "optionYes": "Yes",
-  "optionNo": "No",
-  "legend": "Did the person who died live permanently in England or Wales, or intend to return here?",
+  "question": "At the time of their death did the person who died:",
+  "option1": "live (domicile) permanently in England or Wales",
+  "option2": "intend to return to England or Wales to live (domicile) permanently",
+  "legend": "At the time of their death did the person who died:",
+
   "errors": {
     "domicile": {
       "required": {
-        "summary": "Answer the question about whether the person who died lived permanently in England and Wales",
-        "message": "Select yes if the person who died lived in England and Wales or intended to return here"
+        "summary": "You haven&rsquo;t answered the question about where they lived permanently or planned to live permanently",
+        "message": "Select an option"
       }
     }
   }
