+++ app/resources/en/translation/deceased/newdomicile.json
@@ -1,0 +1,17 @@
+{
+  "title": "Deceased domicile",
+  "question": "Did the person who died live permanently in England or Wales, or intend to return here?",
+  "hintText1": "You can only apply for probate online if the person who died was domiciled in England or Wales at the time of their death.",
+  "hintText2": "You can read more about <a href=\"https://www.gov.uk/inheritance-tax/when-someone-living-outside-the-uk-dies\">domicile</a>.",
+  "optionYes": "Yes",
+  "optionNo": "No",
+  "legend": "Did the person who died live permanently in England or Wales, or intend to return here?",
+  "errors": {
+    "domicile": {
+      "required": {
+        "summary": "Answer the question about whether the person who died lived permanently in England and Wales",
+        "message": "Select yes if the person who died lived in England and Wales or intended to return here"
+      }
+    }
+  }
+}
