+++ app/steps/ui/deceased/domicile/index.js
@@ -1,7 +1,4 @@
-'use strict';
-
 const ValidationStep = require('app/core/steps/ValidationStep');
-const json = require('app/resources/en/translation/deceased/domicile');
 
 module.exports = class DeceasedDomicile extends ValidationStep {
 
@@ -9,16 +6,4 @@ module.exports = class DeceasedDomicile extends ValidationStep {
         return '/deceased-domicile';
     }
 
-    nextStepUrl(ctx) {
-        return this.next(ctx).constructor.getUrl('notInEnglandOrWales');
-    }
-
-    nextStepOptions() {
-        const nextStepOptions = {
-            options: [
-                {key: 'domicile', value: json.optionYes, choice: 'inEnglandOrWales'}
-            ]
-        };
-        return nextStepOptions;
-    }
 };
