+++ app/steps/ui/deceased/newdomicile/schema.json
@@ -1,0 +1,14 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "id": "new-deceased-domicile",
+  "type": "object",
+  "properties": {
+    "domicile": {
+      "type": "string",
+      "minLength": 1
+    }
+  },
+  "required": [
+    "domicile"
+  ]
+}
