+++ test/data/complete-form-multipleapplicants.json
@@ -121,28 +121,28 @@
   },
 
   "declaration": {
-  "legalStatement": {
-    "applicant": "We, Bob Richard Smith of Adam & Eve 81 Petty France London SW1H 9EX and exec_3_new_name of exec_3_address\r\n, make the following statement:",
-    "deceased": "died_firstname died_lastname was born on 1 January 1950 and died on 1 January 2018, domiciled in England and Wales.",
-    "deceasedOtherNames": "They were also known as new_died_firstname new_died_lastname.",
-    "executorsApplying": [
-      {
-        "name": "Bob Richard Smith, an executor named in the will or codicils, is applying for probate.",
-        "sign": "Bob Richard Smith will sign and send to the probate registry what they believe to be the true and original last will and testament of died_firstname died_lastname."
-      },
-      {
-        "name": "exec_3_new_name , an executor named in the will or codicils as executor_3_name, is applying for probate.",
-        "sign": "exec_3_new_name will sign a photocopy of what they believe to be the true and original last will and testament of died_firstname died_lastname. Bob Richard Smith will send the signed photocopy to the probate registry."
-      }
-    ],
-    "deceasedEstateValue": "The gross value for the estate amounts to &pound;123456 and the net value for the estate amounts to &pound;12345.",
-    "deceasedEstateLand": "To the best of our knowledge, information and belief, there was no land vested in died_firstname died_lastname which was settled previously to the death (and not by the will) of died_firstname died_lastname and which remained settled land notwithstanding such death.",
-    "executorsNotApplying": [
-      "executor_2_name, an executor named in the will or codicils, is not making this application because they died before died_firstname died_lastname died.",
-      "executor_4_name, an executor named in the will or codicils, is not making this application now and gives up the right to do so in the future."
-    ],
-    "intro": "This statement is based on the information Bob Richard Smith has given in their application. It will be stored as a public record."
-  },
+    "legalStatement": {
+      "applicant": "We, Bob Richard Smith of Adam & Eve 81 Petty France London SW1H 9EX and exec_3_new_name of exec_3_address\r\n, make the following statement:",
+      "deceased": "died_firstname died_lastname was born on 1 January 1950 and died on 1 January 2018, domiciled in England and Wales.",
+      "deceasedOtherNames": "They were also known as new_died_firstname new_died_lastname.",
+      "executorsApplying": [
+        {
+          "name": "Bob Richard Smith, an executor named in the will or codicils, is applying for probate.",
+          "sign": "Bob Richard Smith will sign and send to the probate registry what they believe to be the true and original last will and testament of died_firstname died_lastname."
+        },
+        {
+          "name": "exec_3_new_name , an executor named in the will or codicils as executor_3_name, is applying for probate.",
+          "sign": "exec_3_new_name will sign a photocopy of what they believe to be the true and original last will and testament of died_firstname died_lastname. Bob Richard Smith will send the signed photocopy to the probate registry."
+        }
+      ],
+      "deceasedEstateValue": "The gross value for the estate amounts to &pound;123456 and the net value for the estate amounts to &pound;12345.",
+      "deceasedEstateLand": "To the best of our knowledge, information and belief, there was no land vested in died_firstname died_lastname which was settled previously to the death (and not by the will) of died_firstname died_lastname and which remained settled land notwithstanding such death.",
+      "executorsNotApplying": [
+        "executor_2_name, an executor named in the will or codicils, is not making this application because they died before died_firstname died_lastname died.",
+        "executor_4_name, an executor named in the will or codicils, is not making this application now and gives up the right to do so in the future."
+      ],
+      "intro": "This statement is based on the information Bob Richard Smith has given in their application. It will be stored as a public record."
+    },
     "declaration": {
       "confirm": "We confirm that we will administer the estate of died_firstname died_lastname, according to law. We will:",
       "confirmItem1": "collect the whole estate",
@@ -156,7 +156,6 @@
       "understandItem2": "criminal proceedings for fraud may be brought against us if we are found to have been deliberately untruthful or dishonest",
       "accept": "I confirm that I will administer the estate of the person who died according to law, and that my application is truthful.",
       "submitWarning": "Once everyone has agreed the legal statement and made their declaration, this information can&rsquo;t be changed."
+    }
+  }
 }
-}
-}
-
