+++ test/data/complete-form.json
@@ -29,7 +29,7 @@
       "dod_date": "2010-03-02T00:00:00.000Z",
       "dod_formattedDate": "1 February 2000",
       "dod_year": 2010,
-      "domicile": "Yes",
+      "domicile": "live permanently in England or Wales",
       "lastName": "Else",
       "postcode": "NW1 8SS",
       "referrer": "DeceasedAddress",
