+++ test/data/deceased.json
@@ -15,7 +15,7 @@
     "dob_month": "12",
     "dob_year": "1990",
     "dob_formattedDate": "12 December 1990",
-    "domicile": "Yes",
+    "domicile": "live permanently in England or Wales",
     "address": "102 Petty France London SW1 1HA"
   }
 }
